+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased](https://github.com/Piturnah/gex/compare/v0.3.3...master)
+## [0.3.4](https://github.com/Piturnah/gex/compare/v0.3.3...v0.3.4) - 2022-12-07
 ### Added
 - Warning when opening Gex with locale other than English ([#13](https://github.com/Piturnah/gex/issues/13))
 - New navigation controls